Linked the menu buttons and set up the win/lose messages
Linked the menu buttons and created and set up the end message for when the player kills the dragon (forest map) or when the player finished the missions in the house map.
Furthermore, I added some flags in the story script to stop the player from going ahead of the story line.